add Python3 support (#6)
I'm merging this for now so people can start using it; I will clean up to make it `six`-ified later.

+++ transformer.py
@@ -37,7 +37,7 @@ class MultiHeadAttention(tf.keras.layers.Layer):
     self.num_heads = num_heads
     self.d_model_size = d_model_size
     
-    self.depth = d_model_size / self.num_heads
+    self.depth = int(d_model_size / self.num_heads)
     
     self.Wq = tf.keras.layers.Dense(d_model_size)
     self.Wk = tf.keras.layers.Dense(d_model_size)
@@ -50,7 +50,7 @@ class MultiHeadAttention(tf.keras.layers.Layer):
     return tf.transpose(x, perm=[0, 2, 1, 3])
     
   def call(self, v, k, q, mask):
-    batch_size = tf.shape(q)[0]
+    batch_size = int(tf.shape(q)[0])
     
     q = self.Wq(q)
     k = self.Wk(k)
